🚚 Rename: 몽뚱그려 분류하던 ModalType을 컴포넌트별로 폴더로 분리하도록 함
