+++ src/pages/SharePage/Components/ShareFolderProfile.css
@@ -1,56 +1,56 @@
 .folder-page-profile {
-	background-color: var(--LBrary-Background);
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	padding-top: 20px;
-	padding-bottom: 60px;
+  background-color: var(--LBrary-Background);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding-top: 20px;
+  padding-bottom: 60px;
 }
 
 .folder-page-profile img {
-	width: 60px;
-	height: 60px;
-	border-radius: 50%;
+  width: 60px;
+  height: 60px;
+  border-radius: 50%;
 }
 
 .folder-page-profile .user-name {
-	margin-top: 12px;
-	font-size: 1.6rem;
-	font-style: normal;
-	font-weight: 400;
-	line-height: 150%;
+  margin-top: 12px;
+  font-size: 1.6rem;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 150%;
 }
 
 .folder-page-profile .folder-name {
-	margin-top: 20px;
-	font-size: 4rem;
-	font-style: normal;
-	font-weight: 600;
-	line-height: normal;
+  margin-top: 20px;
+  font-size: 4rem;
+  font-style: normal;
+  font-weight: 600;
+  line-height: normal;
 }
 
 @media (max-width: 767px) {
-	.folder-page-profile {
-		padding-top: 10px;
-		padding-bottom: 40px;
-	}
+  .folder-page-profile {
+    padding-top: 10px;
+    padding-bottom: 40px;
+  }
 
-	.folder-page-profile img {
-		width: 40px;
-		height: 40px;
-	}
+  .folder-page-profile img {
+    width: 40px;
+    height: 40px;
+  }
 
-	.folder-page-profile .user-name {
-		margin-top: 6px;
-		font-size: 1.4rem;
-		font-weight: 400;
-		line-height: 100%;
-	}
+  .folder-page-profile .user-name {
+    margin-top: 6px;
+    font-size: 1.4rem;
+    font-weight: 400;
+    line-height: 100%;
+  }
 
-	.folder-page-profile .folder-name {
-		margin-top: 10px;
-		font-size: 3.2rem;
-		letter-spacing: -0.02rem;
-	}
+  .folder-page-profile .folder-name {
+    margin-top: 10px;
+    font-size: 3.2rem;
+    letter-spacing: -0.02rem;
+  }
 }
